+++ src/app/feat-recruit/interactions/buttons/cancel_event.ts
@@ -18,7 +18,7 @@ import { StickyKey } from '../../../constant/sticky_key.js';
 import { sendRecruitButtonLog } from '../../../logs/buttons/recruit_button_log.js';
 import { sendErrorLogs } from '../../../logs/error/send_error_logs.js';
 import { RecruitOpCode, regenerateCanvas } from '../../canvases/regenerate_canvas.js';
-import { cancelRecruitEvent } from '../../common/vc_reservation/recruit_event.js';
+import { cancelRecruitEvent } from '../../vc_reservation/recruit_event.js';
 import {
     availableRecruitString,
     getStickyChannelId,
