+++ src/app/feat-recruit/interactions/commands/other_game_recruit.ts
@@ -21,10 +21,10 @@ import { searchMessageById } from '../../../common/manager/message_manager';
 import { assertExistCheck, exists, sleep } from '../../../common/others';
 import { RoleKeySet } from '../../../constant/role_key';
 import { sendErrorLogs } from '../../../logs/error/send_error_logs';
-import { embedRecruitDeleteButton, recruitActionRow } from '../../buttons/create_recruit_buttons';
-import { getVCReserveErrorMessage } from '../../common/create_recruit/condition_checks/vc_reserve_check';
-import { buildOtherGameRecruitText } from '../../common/create_recruit/recruit_text';
-import { createRecruitEvent } from '../../common/vc_reservation/recruit_event';
+import { embedRecruitDeleteButton, recruitActionRow } from '../../create_recruit/common/create_recruit_buttons';
+import { getVCReserveErrorMessage } from '../../create_recruit/condition_checks/vc_reserve_check';
+import { buildOtherGameRecruitText } from '../../create_recruit/common/recruit_text';
+import { createRecruitEvent } from '../../vc_reservation/recruit_event';
 import { sendRecruitSticky } from '../../sticky/recruit_sticky_messages';
 
 const logger = log4js_obj.getLogger('recruit');
