+++ src/app/feat-recruit/sticky/recruit_sticky_messages.ts
@@ -11,7 +11,7 @@ import { sendStickyMessage } from '../../common/sticky_message';
 import { ChannelKeySet } from '../../constant/channel_key';
 import { StickyKey } from '../../constant/sticky_key';
 import { sendErrorLogs } from '../../logs/error/send_error_logs';
-import { createNewRecruitButton } from '../buttons/create_recruit_buttons';
+import { createNewRecruitButton } from '../create_recruit/common/create_recruit_buttons';
 
 const logger = log4js_obj.getLogger('message');
