+++ src/app/handlers/vcState_update_handler.ts
@@ -5,7 +5,7 @@ import { exists, notExists } from '../common/others';
 import { vcToolsStickyFromVoiceState } from '../event/vctools_sticky/vc_tools_message';
 import { disableLimit } from '../event/vctools_sticky/voice_lock';
 import { endCall, startCall } from '../event/voice_count/voice_count';
-import { endEventOnRecruiterLeave } from '../feat-recruit/common/vc_reservation/event_auto_end';
+import { endEventOnRecruiterLeave } from '../feat-recruit/vc_reservation/event_auto_end';
 import { autokill } from '../feat-utils/voice/tts/discordjs_voice';
 import { sendErrorLogs } from '../logs/error/send_error_logs';
